Refactor into new GitHubCheckRunService
Now the methods related to servicing GitHub Check Runs are part of the
GitHubCheckRunService, which is apart from the GitHubRepoService. This
should help make it easier to focus on the check run flow, apart from
the functionality needed to sync data from GitHub.
