+++ src/timessquare/worker/functions/create_rerequested_check_run.py
@@ -8,7 +8,7 @@ from safir.dependencies.db_session import db_session_dependency
 from safir.github.webhooks import GitHubCheckRunEventModel
 from safir.slack.blockkit import SlackCodeBlock, SlackMessage, SlackTextField
 
-from timessquare.worker.servicefactory import create_github_repo_service
+from timessquare.worker.servicefactory import create_github_check_run_service
 
 
 async def create_rerequested_check_run(
@@ -29,7 +29,7 @@ async def create_rerequested_check_run(
 
     try:
         async for db_session in db_session_dependency():
-            github_repo_service = await create_github_repo_service(
+            github_repo_service = await create_github_check_run_service(
                 http_client=ctx["http_client"],
                 logger=logger,
                 installation_id=payload.installation.id,
